backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Werkzeug==3.0.1
google-generativeai>=0.8.0
requests==2.31.0
sendgrid==6.11.0
python-dotenv==1.0.0
python-dateutil==2.8.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
bcrypt==4.1.3

# ── Testing ──────────────────────────────
pytest>=8.0.0
pytest-cov>=4.1.0
flake8>=7.0.0